Extract diagnostics concerns to separate file
